+++ examples/markdown/src/bin/example_markdown_app.rs
@@ -22,7 +22,8 @@ fn main() {
         .flatten()
         .expect_throw("failed to query root element");
 
-    let mut bridge = MarkdownWorker::spawner().spawn("/example_markdown_worker.js");
+    let mut bridge =
+        MarkdownWorker::spawner().spawn_with_loader("/example_markdown_worker_loader.js");
 
     spawn_local(async move {
         let content = bridge.run(MARKDOWN_CONTENT.to_owned()).await;